Fix CRC calculation

+++ CODE/crc.cpp
@@ -62,7 +62,7 @@ void CRCEngine::operator() (char datum)
 {
 	StagingBuffer.Buffer[Index++] = datum;
 
-	if (Index == sizeof(long))  {
+	if (Index == sizeof(int32_t))  {
 		CRC = Value();
 		StagingBuffer.Composite = 0;
 		Index = 0;
@@ -88,7 +88,7 @@ void CRCEngine::operator() (char datum)
  * HISTORY:                                                                                    *
  *   03/02/1996 JLB : Created.                                                                 *
  *=============================================================================================*/
-long CRCEngine::operator() (void const * buffer, int length)
+int32_t CRCEngine::operator() (void const * buffer, int length)
 {
 	if (buffer != NULL && length > 0)  {
 		char const * dataptr = (char const *)buffer;
@@ -110,11 +110,11 @@ long CRCEngine::operator() (void const * buffer, int length)
 		**	Perform the fast 'bulk' processing by reading long word sized
 		**	data blocks.
 		*/
-		long const * longptr = (long const *)dataptr;
-		int longcount = bytes_left / sizeof(long);		// Whole 'long' elements remaining.
+		int32_t const * longptr = (int32_t const *)dataptr;
+		int longcount = bytes_left / sizeof(int32_t);		// Whole 'long' elements remaining.
 		while (longcount--) {
 			CRC = (CRC << 1 | CRC >> 31) + *longptr++;
-			bytes_left -= sizeof(long);
+			bytes_left -= sizeof(int32_t);
 		}
 
 		/*
